requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
django-database-url==1.0.3
Django>=4.2
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.6
tzdata==2022.6
django-cors-headers==4.3.1
requests==2.31.0
channels==4.1.0
uuid==1.30
channels-redis==4.2.0
daphne==4.1.2
onesignal-python-api==2.0.2
sendgrid==6.11.0
python-decouple==3.8
xhtml2pdf==0.2.16
firebase-admin==6.6.0